# interface-specs/openapi/platform-api-v1.yaml
openapi: 3.0.3
info:
  title: 县域系统集成平台管理接口
  version: 0.1.0
  description: >-
    OpenAPI 是用机器可读文件说明 HTTP 接口的通用规范。本文件描述当前平台内部管理接口。
    业务接口只有在接口规范、已删除或遮盖敏感字段的样例、网络条件和负责人均确认后才能开发。
servers:
  - url: /api/v1
security:
  - hospitalIdentity: []
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: 查询主要机构在当前账号可访问范围内的用户
      responses:
        '200':
          description: 当前账号可查看的平台用户
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserListResponse' }
        '403': { description: 'identity:read permission is required' }
    post:
      operationId: createUser
      summary: 在当前账号可管理的机构中创建本地用户
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201': { description: 已创建用户，首次登录必须修改密码 }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { description: 'identity:write permission and organization scope are required' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema: { type: integer, format: int64, minimum: 1 }
    get:
      operationId: getUser
      summary: 查询当前账号有权查看的平台用户
      responses:
        '200':
          description: 用户详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '403': { description: 'identity:read permission and organization scope are required' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
    put:
      operationId: updateUser
      summary: 修改用户显示名称和主要机构
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserRequest' }
      responses:
        '200': { description: 已修改用户 }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { description: 'identity:write permission and both organization scopes are required' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /users/{userId}/enabled:
    patch:
      operationId: changeUserEnabled
      summary: 启用或停用用户，同时保留至少一个可用管理员
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangeEnabledRequest' }
      responses:
        '200': { description: 已修改用户状态 }
        '403': { description: 'identity:write permission and organization scope are required' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /users/{userId}/password-reset:
    post:
      operationId: resetUserPassword
      summary: 设置临时密码并使该用户的现有会话权限失效
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '200': { description: 已设置临时密码 }
        '403': { description: 'identity:write permission and organization scope are required' }
  /users/{userId}/roles:
    put:
      operationId: replaceUserRoles
      summary: 一次性替换用户角色，同时保留至少一个可用管理员
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceRoleIdsRequest' }
      responses:
        '200': { description: 已替换用户角色 }
        '403': { description: 'access:write permission and organization scope are required' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /users/{userId}/organization-scopes:
    put:
      operationId: replaceUserOrganizationScopes
      summary: 在当前操作人可访问的机构内一次性替换用户可访问机构
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceOrganizationIdsRequest' }
      responses:
        '200': { description: 已替换用户可访问机构 }
        '403': { description: 'access:write permission and all organization scopes are required' }
...
components:
  parameters:
    CsrfHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: 由 GET /session/csrf 返回，并与 XSRF-TOKEN Cookie 配套使用的令牌
      schema: { type: string }
  securitySchemes:
    hospitalIdentity:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: >-
        Cookie marker for an authenticated management session. The first development
        phase uses a server-side local-management session and secure HttpOnly cookie;
        production identity must be reviewed by the hospital before activation.
  responses:
    InvalidRequest:
      description: 请求参数校验未通过
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    ResourceNotFound:
      description: 请求的平台数据不存在，或当前账号无权查看
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ResourceConflict:
      description: 唯一值重复、当前状态不允许操作，或数据已被他人修改
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
...
    UserListResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string }
        message: { type: string }
        data:
          type: array
          items: { $ref: '#/components/schemas/ManagedUser' }
    UserResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string }
        message: { type: string }
        data: { $ref: '#/components/schemas/ManagedUser' }
    ManagedUser:
      type: object
      required: [id, loginName, displayName, primaryOrganizationId, organizationCode, enabled, mustChangePassword, createdAt, updatedAt, version, roleIds, organizationScopeIds]
      properties:
        id: { type: integer, format: int64 }
        loginName: { type: string }
        displayName: { type: string }
        primaryOrganizationId: { type: integer, format: int64 }
        organizationCode: { type: string }
        enabled: { type: boolean }
        mustChangePassword: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: string, format: byte }
        roleIds: { type: array, items: { type: integer, format: int64 } }
        organizationScopeIds: { type: array, items: { type: integer, format: int64 } }
    CreateUserRequest:
      type: object
      required: [loginName, displayName, primaryOrganizationId, temporaryPassword]
      properties:
        loginName: { type: string, minLength: 3, maxLength: 64, pattern: '^[A-Za-z0-9._-]+$' }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        primaryOrganizationId: { type: integer, format: int64, minimum: 1 }
        temporaryPassword: { type: string, format: password, writeOnly: true, minLength: 12, maxLength: 128 }
    UpdateUserRequest:
      type: object
      required: [displayName, primaryOrganizationId, version]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 100 }
        primaryOrganizationId: { type: integer, format: int64, minimum: 1 }
        version: { type: string, format: byte }
    ChangeEnabledRequest:
      type: object
      required: [enabled, version]
      properties:
        enabled: { type: boolean }
        version: { type: string, format: byte }
    ResetPasswordRequest:
      type: object
      required: [temporaryPassword]
      properties:
        temporaryPassword: { type: string, format: password, writeOnly: true, minLength: 12, maxLength: 128 }
    ReplaceRoleIdsRequest:
      type: object
      required: [roleIds]
      properties:
        roleIds: { type: array, uniqueItems: true, items: { type: integer, format: int64, minimum: 1 } }
    ReplaceOrganizationIdsRequest:
      type: object
      required: [organizationIds]
      properties:
        organizationIds: { type: array, minItems: 1, uniqueItems: true, items: { type: integer, format: int64, minimum: 1 } }
...
    ApiError:
      type: object
      required: [code, message, timestamp]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }